moved particle code to particle placement, is legacy code now
